Introduced executecore in core which both exe and task can use
