+++ GitVersionCore/GitVersionCore.csproj
@@ -81,6 +81,8 @@
     <Compile Include="Configuration\LegacyConfigNotifier.cs" />
     <Compile Include="Configuration\OldConfigurationException.cs" />
     <Compile Include="EffectiveConfiguration.cs" />
+    <Compile Include="ExecuteCore.cs" />
+    <Compile Include="GitPreparer.cs" />
     <Compile Include="Helpers\FileSystem.cs" />
     <Compile Include="Helpers\IFileSystem.cs" />
     <Compile Include="Helpers\ProcessHelper.cs" />
