+++ GitVersionExe/GitVersionExe.csproj
@@ -56,7 +56,6 @@
     <Compile Include="ArgumentParser.cs" />
     <Compile Include="Arguments.cs" />
     <Compile Include="ExtensionMethods.cs" />
-    <Compile Include="GitPreparer.cs" />
     <Compile Include="HelpWriter.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="AssemblyInfo.cs" />
